Fix: ProductInfo -> ProductDto Inner class로 변경
