--- app/src/main/java/com/codesoom/assignment/application/product/ProductInfo.java
@@ -1,28 +1,0 @@
-package com.codesoom.assignment.application.product;
-
-import com.codesoom.assignment.domain.product.Product;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-@lombok.Generated
-public class ProductInfo {
-    private final Long id;
-
-    private final String name;
-
-    private final String maker;
-
-    private final Long price;
-
-    private final String imageUrl;
-
-    public ProductInfo(Product product) {
-        this.id = product.getId();
-        this.name = product.getName();
-        this.maker = product.getMaker();
-        this.price = product.getPrice();
-        this.imageUrl = product.getImageUrl();
-    }
-}
