+++ app/src/main/java/com/codesoom/assignment/controller/product/ProductDto.java
@@ -1,5 +1,6 @@
 package com.codesoom.assignment.controller.product;
 
+import com.codesoom.assignment.domain.product.Product;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -28,5 +29,28 @@ public class ProductDto {
         private Long price;
 
         private String imageUrl;
+    }
+
+    @Getter
+    @ToString
+    @lombok.Generated
+    public static class ProductInfo {
+        private final Long id;
+
+        private final String name;
+
+        private final String maker;
+
+        private final Long price;
+
+        private final String imageUrl;
+
+        public ProductInfo(Product product) {
+            this.id = product.getId();
+            this.name = product.getName();
+            this.maker = product.getMaker();
+            this.price = product.getPrice();
+            this.imageUrl = product.getImageUrl();
+        }
     }
 }
